BSON `decodeBuffer` support
To support BSON-specific types, we have to construct a `BsonReader` for the buffered data. This patch does this by writing out the data to a byte array and opening a new reader on it. This is a bit wasteful, but seems like the best solution short of building a custom `BsonReader` implementation.