Rework numeric time handling (#432)
This PR reworks the handling for numeric values for parsing and writing java.time classes. The old handling was inconsistent (millis for Instant, ODT, ZDT, but nanos for LT), incomplete (no support for decimals), incompatible with jackson-databind (it uses seconds by default), and not configurable. It also used an "optimistic" decode pattern that broke #425.

These changes remove the old write-dates-as-timestamps setting and replace it with two new settings, time-write-shape and numeric-time-unit. The former controls the format to serialize as (number or string), the latter controls the time unit. The old behavior can be activated with:

```
micronaut.serde.time-write-shape=integer
micronaut.serde.numeric-time-unit=legacy
```

The defaults are to serialize as string by default, removing any unit ambiguity. The unit defaults to seconds, which is the default in jackson-databind.