Test case for private builder constructor method Unfortunately this test case is a bit removed from my original issue.
In the original issue, I have a third-party bean with a builder and jackson annotations. The builder is generated by lombok and has an inaccessible constructor, I think package-private. I want to import that bean with `@SerdeImport`, with the additional qualification of `@Introspected(builder = @Introspected.IntrospectionBuilder(builderMethod = "builder"))` in order to use the builder method instead of the default constructor of the builder class.

The problem is that serde seems to be ignoring my custom `@Introspected` and replacing it with `@Introspected` with `builderClass` instead of `builderMethod`. In the original project, this leads to a runtime illegal access of the builder constructor.

This test case is modified a bunch, because I wanted to restrict it to one module where the processor is already running. Since the class needs to be annotated with `@JsonDeserialize`, and that triggers serde processing, I can't build this test with `@SerdeImport` anymore. I instead moved the `@Introspected` to the actual bean class. I also made the builder class constructor private, to trigger the error: package-private is not enough since the introspection will be placed in the same package.

In the test case, the bug manifests as a compile-time error:

```
/home/yawkat/dev/mn/micronaut-serialization/doc-examples/example-java/src/main/java/example/BuilderBean.java:12: error: No accessible constructor found for builder: example.BuilderBean$Builder
public class BuilderBean {
```