Add ObjectMapper.cloneWithConfiguration This new method allows creating modified ObjectMappers with different configuration. This is required so that we can create an ObjectMapper using the BeanContext in micronaut-oraclecloud, which requires non-default serde configuration.
The changes here are a bit painful. The existing use sites of the config classes all assume the configuration is global and static. They all had to be adjusted.

The API is also painful to use because the configuration classes are interfaces. ConfigCloneSpec shows an example how to use this API without having to implement SerdeConfiguration and friends manually, but it's a bit awkward.