Add ObjectMapper.cloneWithConfiguration (#719)
* Add ObjectMapper.cloneWithConfiguration
This new method allows creating modified ObjectMappers with different configuration. This is required so that we can create an ObjectMapper using the BeanContext in micronaut-oraclecloud, which requires non-default serde configuration.

The changes here are a bit painful. The existing use sites of the config classes all assume the configuration is global and static. They all had to be adjusted.

The API is also painful to use because the configuration classes are interfaces. ConfigCloneSpec shows an example how to use this API without having to implement SerdeConfiguration and friends manually, but it's a bit awkward.

* stuff

* fix

* fix

* address review

* deser test

* move to optional

* Update serde-api/src/main/java/io/micronaut/serde/Deserializer.java

* Update serde-api/src/main/java/io/micronaut/serde/ObjectMapper.java

* Update serde-api/src/main/java/io/micronaut/serde/Deserializer.java

* Update serde-api/src/main/java/io/micronaut/serde/Serializer.java

* Update serde-api/src/main/java/io/micronaut/serde/SerdeRegistry.java

* Update serde-api/src/main/java/io/micronaut/serde/Serializer.java

* inject

---------

Co-authored-by: Sergio del Amo <sergio.delamo@softamo.com>